GPLIM-1304 null check before clearing

+++ mercury/src/main/java/org/broadinstitute/gpinformatics/mercury/presentation/admin/PublicMessageAdminActionBean.java
@@ -62,7 +62,9 @@ public class PublicMessageAdminActionBean extends CoreActionBean {
 
     @HandlesEvent(CLEAR_MESSAGE)
     public Resolution clearMessage() {
-        publicMessageDao.remove(publicMessage);
+        if (publicMessage != null) {
+            publicMessageDao.remove(publicMessage);
+        }
         return getSourcePageResolution();
     }
